requirements.txt
pydub==0.24.1
pyserial==3.4
requests==2.24.0
scipy==1.5.2
